+++ src/mechanisms/scram/server.rs
@@ -1,22 +1,36 @@
 use crate::error::{MechanismError, MechanismErrorKind, SessionError};
 use crate::mechanisms::scram::client::{ProtocolError, SCRAMError};
-use crate::mechanisms::scram::parser::{ClientFinal, ClientFirstMessage, GS2CBindFlag, ServerErrorValue, ServerFinal, ServerFirst};
-use crate::mechanisms::scram::tools::{find_proofs, generate_nonce, DOutput};
+use crate::mechanisms::scram::parser::{
+    ClientFinal, ClientFirstMessage, GS2CBindFlag, ServerErrorValue, ServerFinal, ServerFirst,
+};
+use crate::mechanisms::scram::tools::{compute_signatures, find_proofs, generate_nonce, DOutput};
 use crate::session::{MechanismData, State};
 use crate::vectored_io::VectoredWriter;
 use digest::crypto_common::BlockSizeUser;
 use digest::generic_array::GenericArray;
-use digest::{Digest, OutputSizeUser};
+use digest::{Digest, FixedOutput, OutputSizeUser};
 use hmac::SimpleHmac;
 use rand::{thread_rng, Rng, RngCore};
 use std::io::Write;
 use std::marker::PhantomData;
 use thiserror::Error;
 
-use crate::context::{Demand, DemandReply, Provider, ThisProvider};
+use crate::context::{Demand, DemandReply, Provider};
 use crate::mechanism::Authentication;
-use crate::mechanisms::scram::properties::{HashIterations, PasswordHash, Salt};
+use crate::mechanisms::scram::properties::{Salt, ScramStoredPassword};
 use crate::property::{AuthId, AuthzId};
+
+trait ScramConfig {
+    type DIGEST: Digest + BlockSizeUser;
+    const ALGORITHM_NAME: &'static str;
+
+    const NONCE_LEN: usize;
+
+    const DEFAULT_SALT_LEN: usize = 32;
+    const DEFAULT_ITERATIONS: u32 = 2u32.pow(14);
+
+    const ABORT_IMMEDIATELY: bool = false;
+}
 
 const DEFAULT_ITERATIONS: u32 = 2u32.pow(14); // 16384, TODO check if still reasonable
 const DEFAULT_SALT_LEN: usize = 32;
@@ -39,11 +53,11 @@ impl MechanismError for ScramServerError {
     }
 }
 
-pub struct ScramServer<D: Digest + BlockSizeUser, const N: usize> {
+pub struct ScramServer<D: Digest + BlockSizeUser + FixedOutput, const N: usize> {
     plus: bool,
     state: Option<ScramServerState<D, N>>,
 }
-impl<D: Digest + BlockSizeUser, const N: usize> ScramServer<D, N> {
+impl<D: Digest + BlockSizeUser + FixedOutput, const N: usize> ScramServer<D, N> {
     pub fn new() -> Self {
         Self {
             plus: false,
@@ -68,14 +82,14 @@ impl<const N: usize> WaitingClientFirst<N> {
         Self { nonce: PhantomData }
     }
 
-    pub fn handle_client_first<D: Digest + BlockSizeUser>(
+    pub fn handle_client_first<D: Digest + BlockSizeUser + FixedOutput>(
         self,
         rng: &mut impl Rng,
         session_data: &mut MechanismData,
         client_first: &[u8],
         writer: impl Write,
         written: &mut usize,
-    ) -> Result<WaitingClientFinal<D>, SessionError> {
+    ) -> Result<WaitingClientFinal<N, D>, SessionError> {
         // Step 1: (try to) parse the client message received.
         let client_first @ ClientFirstMessage {
             cbflag,
@@ -84,14 +98,20 @@ impl<const N: usize> WaitingClientFirst<N> {
             nonce: client_nonce,
         } = ClientFirstMessage::parse(client_first).map_err(SCRAMError::ParseError)?;
 
+        // AuthMessage we need to validate the user:
+        // client-first-message-bare + "," + server-first-message + "," + client-final-message-without-proof
+
+        // TODO: Only store this if we're a -PLUS
+        let mut gs2_header = client_first.build_gs2_header_vec();
+
         // FIXME: Escape Username from SCRAM format to whatever
         #[derive(Copy, Clone)]
         struct Prov<'a> {
             authid: &'a str,
             authzid: Option<&'a str>,
         }
-        impl Provider for Prov<'_> {
-            fn provide<'a>(&'a self, req: &mut Demand<'a>) -> DemandReply<()> {
+        impl<'a> Provider<'a> for Prov<'a> {
+            fn provide(&self, req: &mut Demand<'a>) -> DemandReply<()> {
                 req.provide_ref::<AuthId>(self.authid)?;
                 if let Some(authzid) = self.authzid {
                     req.provide_ref::<AuthzId>(authzid)?;
@@ -100,106 +120,122 @@ impl<const N: usize> WaitingClientFirst<N> {
             }
         }
 
+        // TODO: This must at this stage provide so much more info <.<
         let provider = Prov { authid, authzid };
-        let mut gs2_header = client_first.build_gs2_header_vec();
 
         match cbflag {
-            // TODO: check if this is a likely protocol downgrade
-            GS2CBindFlag::SupportedNotUsed => {},
-            GS2CBindFlag::NotSupported => {},
-            GS2CBindFlag::Used(name) => session_data.need_cb_data(name, provider, &mut |cbdata| {
+            // TODO: check if this is a protocol downgrade
+            GS2CBindFlag::SupportedNotUsed => {}
+            GS2CBindFlag::NotSupported => {}
+            GS2CBindFlag::Used(name) => session_data.need_cb_data(name, provider, |cbdata| {
                 gs2_header.extend_from_slice(cbdata);
                 Ok(())
             })?,
         };
 
-        let password: Option<GenericArray<u8, D::OutputSize>>;
-
-        // Retrieve the password for the given user via callback.
-        // If the callback doesn't return a password (usually because the user does not exist) we
-        // proceed with the authentication exchange with randomly generated data, since SCRAM
-        // only indicates failure like that in the last step.
-        password =
-            session_data.maybe_need_with::<PasswordHash, _, _>(&provider, &mut |password| {
-                if password.len() != <SimpleHmac<D> as OutputSizeUser>::output_size() {
+        let params = session_data.maybe_need_with::<ScramStoredPassword, _, _>(
+            &provider,
+            |ScramStoredPassword {
+                 iterations,
+                 salt,
+                 stored_key,
+                 server_key,
+             }| {
+                // First, check if the given values are even possible; we know the digest in
+                // use, we exactly know its output size
+                let hmac_len = <SimpleHmac<D> as OutputSizeUser>::output_size();
+                let hash_len = <D as Digest>::output_size();
+                if stored_key.len() != hash_len || server_key.len() != hmac_len {
                     return Err(SessionError::MechanismError(Box::new(
                         ScramServerError::PasswordHashInvalid,
                     )));
                 }
-                Ok(GenericArray::clone_from_slice(password))
-            })?;
 
-        let (iterations, salt) = if password.is_some() {
-            let iterations = session_data
-                .need_with::<HashIterations, _, _>(&provider, &mut |iterations| Ok(*iterations))?;
-            let salt = session_data
-                .need_with::<Salt, _, _>(&provider, &mut |salt| Ok(base64::encode(salt)))?;
-            (iterations, salt)
-        } else {
-            self.gen_rand_pw_params()
-        };
+                Ok((
+                    format!("{}", iterations),
+                    base64::encode(salt),
+                    GenericArray::clone_from_slice(stored_key),
+                    GenericArray::clone_from_slice(server_key),
+                ))
+            },
+        )?;
 
         let server_nonce: [u8; N] = generate_nonce(rng);
 
-        let it_bytes = format!("{}", iterations);
-        let msg = ServerFirst::new(
-            &client_nonce,
-            &server_nonce,
-            salt.as_bytes(),
-            it_bytes.as_bytes(),
-        );
-        let mut vecw = VectoredWriter::new(msg.to_ioslices());
-        *written = vecw.write_all_vectored(writer)?;
-
-        let mut common_nonce = Vec::with_capacity(client_nonce.len() + server_nonce.len());
-        common_nonce.extend_from_slice(&client_nonce);
-        common_nonce.extend_from_slice(&server_nonce);
-
-        if let Some(salted_password) = password {
-            let gs2_header_b64 = base64::encode(&gs2_header[..]);
-
-            let (proof, signature) = find_proofs::<D>(
-                authid,
-                client_nonce,
-                msg,
-                &gs2_header_b64,
-                &GenericArray::from_slice(&salted_password),
+        if let Some((iterations, salt, stored_key, server_key)) = params {
+            let msg = ServerFirst::new(
+                &client_nonce,
+                &server_nonce,
+                salt.as_bytes(),
+                iterations.as_bytes(),
             );
-            Ok(WaitingClientFinal::new(common_nonce, gs2_header, proof, signature))
+            let mut vecw = VectoredWriter::new(msg.to_ioslices());
+            *written = vecw.write_all_vectored(writer)?;
+
+            Ok(WaitingClientFinal::new(
+                client_nonce.into(),
+                server_nonce,
+                gs2_header,
+                authid.to_string(),
+                salt,
+                iterations,
+                stored_key,
+                server_key,
+            ))
         } else {
-            Ok(WaitingClientFinal::bad_user(common_nonce, gs2_header))
+            let mut salt = [0u8; DEFAULT_SALT_LEN];
+            thread_rng().fill_bytes(&mut salt);
+            let salt = base64::encode(salt);
+
+            let msg = ServerFirst::new(&client_nonce, &server_nonce, salt.as_bytes(), b"16384");
+            let mut vecw = VectoredWriter::new(msg.to_ioslices());
+            *written = vecw.write_all_vectored(writer)?;
+
+            Ok(WaitingClientFinal::bad_user())
         }
-    }
-
-    fn gen_rand_pw_params(&self) -> (u32, String) {
-        let mut salt = [0u8; DEFAULT_SALT_LEN];
-        thread_rng().fill_bytes(&mut salt);
-
-        (DEFAULT_ITERATIONS, base64::encode(salt))
     }
 }
 
-pub struct WaitingClientFinal<D: Digest + BlockSizeUser> {
-    nonce: Vec<u8>,
+pub struct WaitingClientFinal<const N: usize, D: Digest + BlockSizeUser + FixedOutput> {
+    data: Option<FinalInner<N, D>>,
+}
+struct FinalInner<const N: usize, D: Digest + BlockSizeUser + FixedOutput> {
+    client_nonce: Vec<u8>,
+    server_nonce: [u8; N],
     gs2_header: Vec<u8>,
-    data: Option<FinalInner<D>>,
+    username: String,
+    salt: String,
+    iterations: String,
+    stored_key: GenericArray<u8, D::OutputSize>,
+    server_key: DOutput<D>,
 }
-struct FinalInner<D: Digest + BlockSizeUser> {
-    proof: DOutput<D>,
-    signature: DOutput<D>,
-}
-impl<D: Digest + BlockSizeUser> WaitingClientFinal<D> {
-    pub fn new(nonce: Vec<u8>, gs2_header: Vec<u8>, proof: DOutput<D>, signature: DOutput<D>)
-        -> Self
-    {
+impl<const N: usize, D: Digest + BlockSizeUser + FixedOutput> WaitingClientFinal<N, D> {
+    pub fn new(
+        client_nonce: Vec<u8>,
+        server_nonce: [u8; N],
+        gs2_header: Vec<u8>,
+        username: String,
+        salt: String,
+        iterations: String,
+        stored_key: GenericArray<u8, D::OutputSize>,
+        server_key: DOutput<D>,
+    ) -> Self {
         Self {
-            nonce,
-            gs2_header,
-            data: Some(FinalInner { proof, signature }),
+            data: Some(FinalInner {
+                client_nonce,
+                server_nonce,
+                gs2_header,
+                username,
+                salt,
+                iterations,
+                stored_key,
+                server_key,
+            }),
         }
     }
-    pub fn bad_user(nonce: Vec<u8>, gs2_header: Vec<u8>) -> Self {
-        Self { nonce, gs2_header, data: None }
+
+    pub fn bad_user() -> Self {
+        Self { data: None }
     }
 
     pub fn handle_client_final(
@@ -207,55 +243,93 @@ impl<D: Digest + BlockSizeUser> WaitingClientFinal<D> {
         client_final: &[u8],
         writer: impl Write,
         written: &mut usize,
-    ) -> Result<Outcome, SessionError> {
+    ) -> Result<(), SessionError> {
         let ClientFinal {
             channel_binding,
             nonce,
             proof,
         } = ClientFinal::parse(client_final).map_err(SCRAMError::ParseError)?;
 
-        let outcome = Outcome::Failed;
+        let msg = if let Some(FinalInner {
+            client_nonce,
+            server_nonce,
+            gs2_header,
+            username,
+            salt,
+            iterations,
+            stored_key,
+            server_key,
+        }) = self.data
+        {
+            let cb = base64::decode(channel_binding)
+                .map_err(|_| SCRAMError::Protocol(ProtocolError::Base64Decode))?;
 
-        let cb = base64::decode(channel_binding)
-            .map_err(|_| SCRAMError::Protocol(ProtocolError::Base64Decode))?;
-
-
-        let msg = if !self.verify_channel_bindings(&cb) {
-            ServerFinal::Error(ServerErrorValue::ChannelBindingsDontMatch)
-        } else {
-            if nonce != self.nonce {
-                ServerFinal::Error(ServerErrorValue::InvalidProof)
+            if &gs2_header[..] != &cb[..] {
+                ServerFinal::Error(ServerErrorValue::ChannelBindingsDontMatch)
             } else {
-                if let Some(data) = self.data {
-                    if proof != data.proof.as_slice() {
-                        ServerFinal::Error(ServerErrorValue::InvalidProof)
+                if let Some(remainder) = nonce.strip_prefix(&client_nonce[..]) {
+                    if remainder == server_nonce {
+                        if proof.len() > (<SimpleHmac<D> as OutputSizeUser>::output_size() * 4 / 3) + 3 {
+                            ServerFinal::Error(ServerErrorValue::InvalidProof)
+                        } else {
+                            let mut proof_decoded = DOutput::<D>::default();
+                            base64::decode_config_slice(proof, base64::STANDARD, &mut proof_decoded).map_err(|_| SCRAMError::Protocol(ProtocolError::Base64Decode));
+
+                            let mut client_signature = DOutput::<D>::default();
+                            let mut server_signature = DOutput::<D>::default();
+
+                            compute_signatures::<D>(
+                                &stored_key,
+                                &server_key,
+                                &username,
+                                &client_nonce,
+                                &server_nonce,
+                                salt.as_bytes(),
+                                iterations.as_bytes(),
+                                channel_binding,
+                                &mut client_signature,
+                                &mut server_signature,
+                            );
+
+                            // Calculate the client_key by XORing the provided proof with the
+                            // calculated client signature
+                            let client_key = DOutput::<D>::from_exact_iter(
+                                proof_decoded
+                                    .into_iter()
+                                    .zip(client_signature)
+                                    .map(|(x, y)| x ^ y),
+                            )
+                                .expect("XOR of two same-sized arrays was not of that size?");
+
+                            let calculated_stored_key = D::digest(client_key);
+
+                            if stored_key != calculated_stored_key {
+                                ServerFinal::Error(ServerErrorValue::InvalidProof)
+                            } else {
+                                let encoded = base64::encode(server_signature);
+                                let msg = ServerFinal::Verifier(encoded.as_bytes());
+                                let mut vecw = VectoredWriter::new(msg.to_ioslices());
+                                *written = vecw.write_all_vectored(writer)?;
+
+                                return Ok(());
+                            }
+                        }
                     } else {
-                        let msg = ServerFinal::Verifier(data.signature.as_slice());
-                        let mut vecw = VectoredWriter::new(msg.to_ioslices());
-                        * written = vecw.write_all_vectored(writer)?;
-                        // FIXME: We need to validate the authzid/authid combo first!
-                        return Ok(Outcome::Successful { username: () });
+                        ServerFinal::Error(ServerErrorValue::InvalidProof)
                     }
                 } else {
-                    ServerFinal::Error(ServerErrorValue::UnknownUser)
+                    ServerFinal::Error(ServerErrorValue::InvalidProof)
                 }
             }
+        } else {
+            ServerFinal::Error(ServerErrorValue::UnknownUser)
         };
 
         let mut vecw = VectoredWriter::new(msg.to_ioslices());
         *written = vecw.write_all_vectored(writer)?;
 
-        Ok(outcome)
+        Ok(())
     }
-
-    pub fn verify_channel_bindings(&self, channel_binding: &[u8]) -> bool {
-        &self.gs2_header[..] == channel_binding
-    }
-}
-
-pub enum Outcome {
-    Failed,
-    Successful { username: () },
 }
 
 struct ScramState<S> {
@@ -268,39 +342,39 @@ impl<const N: usize> ScramState<WaitingClientFirst<N>> {
         }
     }
 
-    pub fn step<D: Digest + BlockSizeUser>(
+    pub fn step<D: Digest + BlockSizeUser + FixedOutput>(
         self,
         rng: &mut impl Rng,
         session_data: &mut MechanismData,
         input: &[u8],
         writer: impl Write,
         written: &mut usize,
-    ) -> Result<ScramState<WaitingClientFinal<D>>, SessionError> {
+    ) -> Result<ScramState<WaitingClientFinal<N, D>>, SessionError> {
         let state = self
             .state
             .handle_client_first(rng, session_data, input, writer, written)?;
         Ok(ScramState { state })
     }
 }
-impl<D: Digest + BlockSizeUser> ScramState<WaitingClientFinal<D>> {
+impl<const N: usize, D: Digest + BlockSizeUser + FixedOutput> ScramState<WaitingClientFinal<N, D>> {
     pub fn step(
         self,
         input: &[u8],
         writer: impl Write,
         written: &mut usize,
-    ) -> Result<ScramState<Outcome>, SessionError> {
+    ) -> Result<ScramState<()>, SessionError> {
         let state = self.state.handle_client_final(input, writer, written)?;
         Ok(ScramState { state })
     }
 }
 
-enum ScramServerState<D: Digest + BlockSizeUser, const N: usize> {
+enum ScramServerState<D: Digest + BlockSizeUser + FixedOutput, const N: usize> {
     WaitingClientFirst(ScramState<WaitingClientFirst<N>>),
-    WaitingClientFinal(ScramState<WaitingClientFinal<D>>),
-    Finished(ScramState<Outcome>),
+    WaitingClientFinal(ScramState<WaitingClientFinal<N, D>>),
+    Finished(ScramState<()>),
 }
 
-impl<D: Digest + BlockSizeUser, const N: usize> Authentication for ScramServer<D, N> {
+impl<D: Digest + BlockSizeUser + FixedOutput, const N: usize> Authentication for ScramServer<D, N> {
     fn step(
         &mut self,
         session: &mut MechanismData,